+++ yaml-config-updater/src/main/java/ru/vyarus/yaml/updater/parse/comments/CommentsReader.java
@@ -13,6 +13,7 @@ import java.text.CharacterIterator;
 import java.text.StringCharacterIterator;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -29,6 +30,9 @@ import java.util.List;
  * @since 22.04.2021
  */
 public final class CommentsReader {
+
+    // assume explicit sequence and object notion as non parsable (keep and store as-is)
+    public static final List<Character> VALUE_BOUNDARY_START = Arrays.asList('\'', '"', '{', '[');
 
     private CommentsReader() {
     }
@@ -124,7 +128,7 @@ public final class CommentsReader {
 
                 // property-like structure might be quoted (simple string)
                 Prop lprop = null;
-                if (chars.current() != '\'' && chars.current() != '"') {
+                if (!VALUE_BOUNDARY_START.contains(chars.current())) {
                     lprop = parseProperty(chars, line);
                 }
                 if (lprop == null) {
@@ -258,6 +262,24 @@ public final class CommentsReader {
                 return true;
             } else if (multiline != null) {
                 // obviously multiline ended
+
+                // here we look for the last lines - it could be empty lines appended by mistake
+                // (ending = 1 is a '+' case - greedy newline "eating" all whitespace after it)
+                if (multiline.ending != 1) {
+                    // in all other cases, whitespace lines must be re-considered as comments
+                    final List<String> lines = current.getValue();
+                    for(int i = lines.size() - 1; i > 0; i--) {
+                        final String ln = lines.get(i);
+                        // detaching empty ending line (but only with lesser indent)
+                        if (ln.trim().isEmpty() && ln.length() < multiline.indent) {
+                            // insert at 0 because we go backward
+                            comments.add(0, lines.remove(i));
+                            continue;
+                        }
+                        break;
+                    }
+                }
+
                 multiline = null;
             }
             return false;